Add files via upload
This repository contains a digit classification project implemented using a neural network. The project is built on the popular MNIST dataset, which consists of a large collection of handwritten digits. The neural network model is trained on this dataset to accurately classify and predict the digits.
The main objective of this project is to showcase the application of neural networks in solving image classification problems. By using the MNIST dataset, which is widely used as a benchmark in the field of machine learning, the project demonstrates the effectiveness of neural networks in recognizing and distinguishing different handwritten digits.